Added ViewAssetMetadata
