+++ dex-demo/client/src/components/ViewAsset.tsx
@@ -5,185 +5,12 @@ import { useAuth } from "../contexts/AuthContext.js";
 import { useApi } from "../contexts/ApiContext.js";
 import {
     Box,
-    Button,
     Tab,
     Tabs,
-    Table,
-    TableBody,
-    TableContainer,
-    TableCell,
-    TableRow,
-    TextField,
     Typography
 } from "@mui/material";
 
-function ViewAssetMetadata({ asset, onSave }: { asset: any, onSave: () => void }) {
-    const { did } = useParams();
-    const navigate = useNavigate();
-    const auth = useAuth();
-    const api = useApi();
-    const { showSnackbar } = useSnackbar();
-    const [newTitle, setNewTitle] = useState<string>(asset.title || "");
-    const [currentTitle, setCurrentTitle] = useState<string>(asset.title || "");
-    const [firstDid, setFirstDid] = useState<string | null>(null);
-    const [prevDid, setPrevDid] = useState<string | null>(null);
-    const [nextDid, setNextDid] = useState<string | null>(null);
-    const [lastDid, setLastDid] = useState<string | null>(null);
-
-    function findAdjacentDids(list: string[], targetDid: string) {
-        let prevDid = null;
-        let nextDid = null;
-        let firstDid = null;
-        let lastDid = null;
-
-        for (let i = 0; i < list.length; i++) {
-            if (list[i] === targetDid) {
-                if (i > 0) {
-                    prevDid = list[i - 1];
-                }
-                if (i < list.length - 1) {
-                    nextDid = list[i + 1];
-                }
-                break;
-            }
-        }
-
-        firstDid = list[0];
-
-        if (firstDid === targetDid) {
-            firstDid = null;
-        }
-
-        lastDid = list[list.length - 1];
-
-        if (lastDid === targetDid) {
-            lastDid = null;
-        }
-
-        return { firstDid, prevDid, nextDid, lastDid };
-    }
-
-    useEffect(() => {
-        const init = async () => {
-            const { firstDid, prevDid, nextDid, lastDid } = findAdjacentDids(asset.collection.assets, asset.did);
-
-            setFirstDid(firstDid);
-            setPrevDid(prevDid);
-            setNextDid(nextDid);
-            setLastDid(lastDid);
-        };
-
-        init();
-    }, [asset]);
-
-    async function saveTitle() {
-        try {
-            const title = newTitle.trim();
-            await api.patch(`/asset/${did}`, { title });
-            setNewTitle(title);
-            setCurrentTitle(title);
-            onSave();
-        }
-        catch (error: any) {
-            showSnackbar("Failed to set profile name", 'error');
-        }
-    }
-
-    return (
-        <TableContainer>
-            <Table>
-                <TableBody>
-                    <TableRow>
-                        <TableCell>Title</TableCell>
-                        {auth.isAuthenticated && auth.userDID === asset.tokenized.owner ? (
-                            <Box sx={{ display: 'flex', alignItems: 'center', gap: 3 }}>
-                                <TextField
-                                    label=""
-                                    value={newTitle}
-                                    onChange={(e) => setNewTitle(e.target.value)}
-                                    slotProps={{
-                                        htmlInput: {
-                                            maxLength: 20,
-                                        },
-                                    }}
-                                    sx={{ width: 300 }}
-                                    margin="normal"
-                                    fullWidth
-                                />
-                                <Button
-                                    variant="contained"
-                                    color="primary"
-                                    onClick={saveTitle}
-                                    disabled={newTitle === currentTitle}
-                                >
-                                    Save
-                                </Button>
-                            </Box>
-                        ) : (
-                            <TableCell>{asset.title || 'no title'}</TableCell>
-                        )}
-                    </TableRow>
-                    <TableRow>
-                        <TableCell>Collection</TableCell>
-                        {asset.collection?.name ? (
-                            <TableCell>
-                                <div style={{ display: 'inline-block' }}>
-                                    <Button
-                                        color="inherit"
-                                        disabled={!firstDid}
-                                        onClick={() => navigate(`/asset/${firstDid}`)}>
-                                        {'<<'}
-                                    </Button>
-                                    <Button
-                                        color="inherit"
-                                        disabled={!prevDid}
-                                        onClick={() => navigate(`/asset/${prevDid}`)}>
-                                        {'<'}
-                                    </Button>
-                                    <a href={`/collection/${asset.collection.did}`}>{asset.collection.name}</a>
-                                    <Button
-                                        color="inherit"
-                                        disabled={!nextDid}
-                                        onClick={() => navigate(`/asset/${nextDid}`)}>
-                                        {'>'}
-                                    </Button>
-                                    <Button
-                                        color="inherit"
-                                        disabled={!lastDid}
-                                        onClick={() => navigate(`/asset/${lastDid}`)}>
-                                        {'>>'}
-                                    </Button>
-                                </div>
-                            </TableCell>
-                        ) : (
-                            <TableCell>no collection</TableCell>
-                        )}
-                    </TableRow>
-                    <TableRow>
-                        <TableCell>Creator</TableCell>
-                        {asset.owner?.name ? (
-                            <TableCell><a href={`/profile/${asset.tokenized.owner}`}>{asset.owner.name}</a></TableCell>
-                        ) : (
-                            <TableCell>no creator</TableCell>
-                        )}
-                    </TableRow>
-                    <TableRow>
-                        <TableCell>File size</TableCell>
-                        <TableCell>{asset.image.bytes} bytes</TableCell>
-                    </TableRow>
-                    <TableRow>
-                        <TableCell>Image size</TableCell>
-                        <TableCell>{asset.image.width} x {asset.image.height} pixels</TableCell>
-                    </TableRow>
-                    <TableRow>
-                        <TableCell>Image type</TableCell>
-                        <TableCell>{asset.image.type}</TableCell>
-                    </TableRow>
-                </TableBody>
-            </Table>
-        </TableContainer>
-    );
-}
+import ViewAssetMetadata from "./ViewAssetMetadata.js";
 
 function ViewAsset() {
     const { did } = useParams();
